Retira ApplicantionController como dependência de serviço
