+++ src/main/java/org/fogbowcloud/manager/core/controllers/ApplicationController.java
@@ -102,7 +102,15 @@ public class ApplicationController {
 	}
 
 	public void createOrder(Order order, String accessId, String localTokenId) throws OrdersServiceException, UnauthorizedException {
-		this.ordersService.createOrder(order, accessId, localTokenId);
+		Token federatedToken = authenticate(accessId);
+		Token localToken = createLocalToken(localTokenId);
+		this.ordersService.createOrder(order, federatedToken, localToken);
+	}
+
+	private Token createLocalToken(String localTokenId) {
+		Token localToken = new Token();
+		localToken.setAccessId(localTokenId);
+		return localToken;
 	}
 
 }
